+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 - Selftests are now run as a pre-commit hook
+- A dummy lab and board for running selftests
+  that are as generic as possible
 
 
 ## [0.3.0] - 2018-06-15