+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - A dummy lab and board for running selftests
   that are as generic as possible
 
+### Changed
+- git testcases now have a `rev` parameter for checking out a specific
+  revision
+
 
 ## [0.3.0] - 2018-06-15
 ### Added